Enrich DB insight definitions with dimensions and metrics (#200)
* Add rawDimensionToDimension, rawMetricToMetric to DBInsightDefinition

The "DB" insight definition type is used for queries against chwww
(and potentially other databases in the future). However, it doesn't
currently define a way to specify which columns in the executed query
correspond to particular dimensions or metrics.

For the `security.device.firmware` insight driving the development of
this feature, I need at least the `device_id` dimension. Metric is less
obviously justifiable, but it's cheaper to add it and not need it than
to miss it out.

This approach is copied from the `KSQLInsightDefinition`, where it
works well. `KDEInsightDefinition` just uses `dimensions` and `metrics`
but it's useful to be able to rename the SQL results.

GitOrigin-RevId: b7d88398b7a5b5acf9a7e8ac847bb8de104403d2